requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple/
asgiref==3.2.10
boto3==1.14.51
botocore==1.17.51
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2.1
django-storages==1.9.1
django==3.1.13
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.5
